// packages/perennial-extensions/types/generated/@equilibria/perennial-v2-oracle/contracts/IOracleProvider.ts
export type OracleVersionStruct = {
  version: PromiseOrValue<BigNumberish>;
  timestamp: PromiseOrValue<BigNumberish>;
  price: PromiseOrValue<BigNumberish>;
  valid: PromiseOrValue<boolean>;
};

export type OracleVersionStructOutput = [
  BigNumber,
  BigNumber,
  BigNumber,
  boolean
] & {
  version: BigNumber;
  timestamp: BigNumber;
  price: BigNumber;
  valid: boolean;
};

export interface IOracleProviderInterface extends utils.Interface {
  functions: {
    "at(uint256)": FunctionFragment;
    "current()": FunctionFragment;
    "latest()": FunctionFragment;
    "sync()": FunctionFragment;
  };

  getFunction(
    nameOrSignatureOrTopic: "at" | "current" | "latest" | "sync"
  ): FunctionFragment;

  encodeFunctionData(
    functionFragment: "at",
    values: [PromiseOrValue<BigNumberish>]
  ): string;
  encodeFunctionData(functionFragment: "current", values?: undefined): string;
  encodeFunctionData(functionFragment: "latest", values?: undefined): string;
  encodeFunctionData(functionFragment: "sync", values?: undefined): string;

  decodeFunctionResult(functionFragment: "at", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "current", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "latest", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "sync", data: BytesLike): Result;

  events: {};
}

export interface IOracleProvider extends BaseContract {
  connect(signerOrProvider: Signer | Provider | string): this;
  attach(addressOrName: string): this;
  deployed(): Promise<this>;

  interface: IOracleProviderInterface;

  queryFilter<TEvent extends TypedEvent>(
    event: TypedEventFilter<TEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TEvent>>;

  listeners<TEvent extends TypedEvent>(
    eventFilter?: TypedEventFilter<TEvent>
  ): Array<TypedListener<TEvent>>;
  listeners(eventName?: string): Array<Listener>;
  removeAllListeners<TEvent extends TypedEvent>(
    eventFilter: TypedEventFilter<TEvent>
  ): this;
  removeAllListeners(eventName?: string): this;
  off: OnEvent<this>;
  on: OnEvent<this>;
  once: OnEvent<this>;
  removeListener: OnEvent<this>;

  functions: {
    at(
      version: PromiseOrValue<BigNumberish>,
      overrides?: CallOverrides
    ): Promise<[OracleVersionStructOutput]>;

    current(overrides?: CallOverrides): Promise<[BigNumber]>;

    latest(overrides?: CallOverrides): Promise<[OracleVersionStructOutput]>;

    sync(
      overrides?: Overrides & { from?: PromiseOrValue<string> }
    ): Promise<ContractTransaction>;
  };

  at(
    version: PromiseOrValue<BigNumberish>,
    overrides?: CallOverrides
  ): Promise<OracleVersionStructOutput>;

  current(overrides?: CallOverrides): Promise<BigNumber>;

  latest(overrides?: CallOverrides): Promise<OracleVersionStructOutput>;

  sync(
    overrides?: Overrides & { from?: PromiseOrValue<string> }
  ): Promise<ContractTransaction>;

  callStatic: {
    at(
      version: PromiseOrValue<BigNumberish>,
      overrides?: CallOverrides
    ): Promise<OracleVersionStructOutput>;

    current(overrides?: CallOverrides): Promise<BigNumber>;

    latest(overrides?: CallOverrides): Promise<OracleVersionStructOutput>;

    sync(
      overrides?: CallOverrides
    ): Promise<[OracleVersionStructOutput, BigNumber]>;
  };

  filters: {};

  estimateGas: {
    at(
      version: PromiseOrValue<BigNumberish>,
      overrides?: CallOverrides
    ): Promise<BigNumber>;

    current(overrides?: CallOverrides): Promise<BigNumber>;

    latest(overrides?: CallOverrides): Promise<BigNumber>;

    sync(
      overrides?: Overrides & { from?: PromiseOrValue<string> }
    ): Promise<BigNumber>;
  };

  populateTransaction: {
    at(
      version: PromiseOrValue<BigNumberish>,
      overrides?: CallOverrides
    ): Promise<PopulatedTransaction>;

    current(overrides?: CallOverrides): Promise<PopulatedTransaction>;

    latest(overrides?: CallOverrides): Promise<PopulatedTransaction>;

    sync(
      overrides?: Overrides & { from?: PromiseOrValue<string> }
    ): Promise<PopulatedTransaction>;
  };
}